Edit & Update migrations, seeder, admin navigation

+++ database/migrations/2024_02_17_184527_create_images_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('images', function (Blueprint $table) {
             $table->id();
-            $table->tinyInteger('ranking');
+            $table->smallInteger('ranking');
             $table
                 ->foreignId('page_id')
                 ->constrained()
